Upgrade libjpeg to version 7.
As I was starting to attempt to compile leptonica and I ran into gcc not being
able to find stdint.h. I noticed that the busybox patch included took all
stdint.h includes out and re-added all the typedefs that were needed. I'm
wondering why this is the case. Why with the default installation can't we
include stdint.h (among other headers it seems)?

In any case I got parts of it to compile :). Now I need to compile some of the
external dependencies (libtiff, libpng, libnetpbm, etc.). I started compiling
libjpeg (in the upstream directory) and it looks like there is a newer version:
http://www.ijg.org/files/jpegsrc.v7.tar.gz which unfortunately fails to compile.
Here's a start at fixing it: